Revert "update for the review"
This reverts commit 062279188a1bd3cb5d917acfef3ed20f00a7f619.
