+++ docs/examples/color/colors.py
@@ -8,13 +8,13 @@ from compas_viewer import Viewer
 viewer = Viewer()
 
 box = Box(1, 1, 1, Frame((0, 0, 0), [1, 0, 0], [0, 1, 0]))
-obj1 = viewer.scene.add(box, facescolor=Color(1.0, 0.0, 0.0), opacity=0.7)
+obj1 = viewer.add(box, facescolor=Color(1.0, 0.0, 0.0), opacity=0.7)
 
 box = Box(1, 1, 1, Frame((0, 0, 0), [1, 0, 0], [0, 1, 0]))
-obj2 = viewer.scene.add(box, facescolor=Color(0.0, 1.0, 0.0), opacity=0.7)
+obj2 = viewer.add(box, facescolor=Color(0.0, 1.0, 0.0), opacity=0.7)
 
 box = Box(1, 1, 1, Frame((0, 0, 0), [1, 0, 0], [0, 1, 0]))
-obj3 = viewer.scene.add(box, facescolor=Color(0.0, 0.0, 1.0), opacity=0.7)
+obj3 = viewer.add(box, facescolor=Color(0.0, 0.0, 1.0), opacity=0.7)
 
 
 obj2.transformation = Translation.from_vector([2, 0, 0])
