+++ docs/examples/dynamic/dynamic.py
@@ -5,7 +5,7 @@ from compas_viewer import Viewer
 from compas_viewer.scene import PointObject
 
 viewer = Viewer()
-obj: PointObject = viewer.scene.add(Point(0, 0, 0), show_points=True, pointscolor=Color.red(), pointsize=10)  # type: ignore
+obj: PointObject = viewer.add(Point(0, 0, 0), show_points=True, pointscolor=Color.red(), pointsize=10)  # type: ignore
 
 
 @viewer.on(interval=1000)
