+++ docs/examples/dynamic/orbiting.py
@@ -15,7 +15,7 @@ viewer.renderer.camera.rotation.x = 10
 viewer.renderer.camera.rotation.y = 10
 viewer.renderer.camera.distance = 5
 
-viewer.scene.add(sphere, facescolor=Color.cyan(), linescolor=Color.blue())
+viewer.add(sphere, facescolor=Color.cyan(), linescolor=Color.blue())
 
 
 @viewer.on(interval=50)
