+++ docs/examples/layout/tree.py
@@ -9,10 +9,10 @@ viewer = Viewer(rendermode="shaded")
 viewer = Viewer()
 for i in range(10):
     for j in range(10):
-        sp = viewer.scene.add(Sphere(1, Frame([10*i, 10*j, 0], [1, 0, 0], [0, 1, 0])), name=f"Sphere_{i}_{j}")
+        sp = viewer.add(Sphere(1, Frame([10*i, 10*j, 0], [1, 0, 0], [0, 1, 0])), name=f"Sphere_{i}_{j}")
 
 
-viewer.layout.sidedock.add_element(Treeform(viewer.scene.tree, {"Name": ".object.name", "Object": ".object"}))
+viewer.layout.sidedock.add_element(Treeform(viewer._tree, {"Name": ".object.name", "Object": ".object"}))
 
 
 viewer.show()
