+++ docs/examples/object/points.py
@@ -8,7 +8,6 @@ from compas_viewer import Viewer
 viewer = Viewer()
 for i in range(10):
     point = Point(random() * 10, random() * 10, random() * 10)
-    viewer.scene.add(point, pointscolor=Color(random(), random(), random()), pointssize=30)
-
+    viewer.add(point, pointscolor=Color(random(), random(), random()), pointssize=30, show_points=True)
 
 viewer.show()
