+++ docs/examples/object/texts.py
@@ -5,12 +5,12 @@ from compas_viewer import Viewer
 viewer = Viewer()
 
 t = Tag("a", (0, 0, 0), height=50)
-viewer.scene.add(t)
+viewer.add(t)
 
 t = Tag("123", (3, 0, 0), height=50)
-viewer.scene.add(t)
+viewer.add(t)
 
 t = Tag("ABC", (3, 3, 0), height=20, absolute_height=True)
-viewer.scene.add(t, color=(1, 0, 0))
+viewer.add(t, color=(1, 0, 0))
 
 viewer.show()
