+++ docs/examples/object/vector.py
@@ -14,6 +14,6 @@ for i in range(0, 360, 20):
     for j in range(0, 180, 10):
         position = Vector(sin(radians(i)) * sin(radians(j)), cos(radians(i)) * sin(radians(j)), cos(radians(j)))
         vector = Vector(sin(radians(i)), cos(radians(i)), cos(radians(j)))
-        viewer.scene.add(vector, anchor=position, linescolor=Color(random(), random(), random()))
+        viewer.add(vector, anchor=position, linescolor=Color(random(), random(), random()))
 
 viewer.show()
