+++ src/compas_viewer/actions/delete_selected.py
@@ -5,8 +5,8 @@ class DeleteSelected(Action):
     """Permanently delete selected objects from the viewer and release the memory."""
 
     def pressed_action(self):
-        for obj in self.scene.objects:
+        for obj in self.viewer.objects:
             if obj.is_selected:
-                self.scene.remove(obj)
+                self.viewer.remove(obj)
                 del obj
         self.viewer.renderer.update()
