+++ src/compas_viewer/actions/select_all.py
@@ -5,7 +5,6 @@ class SelectAll(Action):
     """Select all objects."""
 
     def pressed_action(self):
-        for obj in self.scene.objects:
-            if obj.is_visible and not obj.is_locked:
-                obj.is_selected = True
+        for obj in self.viewer.objects:
+            obj.is_selected = True
         self.viewer.renderer.update()
