+++ src/compas_viewer/viewer.py
@@ -1,9 +1,15 @@
 import sys
 from pathlib import Path
+from typing import Any
 from typing import Callable
 from typing import Literal
 from typing import Optional
+from typing import Union
 
+from compas.colors import Color
+from compas.data import Data
+from compas.geometry import Frame
+from compas.scene import Scene
 from PySide6.QtCore import QCoreApplication
 from PySide6.QtWidgets import QApplication
 from PySide6.QtWidgets import QMainWindow
@@ -18,14 +24,15 @@ from compas_viewer.configurations import RendererConfig
 from compas_viewer.configurations import SceneConfig
 from compas_viewer.controller import Controller
 from compas_viewer.layout import Layout
+from compas_viewer.scene import FrameObject
+from compas_viewer.scene import ViewerSceneObject
 from compas_viewer.utilities import Timer
+from compas_viewer.utilities import instance_colors_generator
 
-from .scene import ViewerScene as Scene
 
-
-class Viewer:
+class Viewer(Scene):
     """
-    The Viewer class is the main entry of the viewer. It contains the scene and create the GUI application.
+    The Viewer class is the main entry of `compas_viewer`. It organizes the scene and create the GUI application.
 
     Parameters
     ----------
@@ -49,8 +56,6 @@ class Viewer:
 
     Attributes
     ----------
-    scene : :class:`compas_viewer.scene.ViewerScene`
-        The scene of the viewer.
     render : :class:`compas_viewer.components.render.Render`
         The render component of the viewer.
     controller : :class:`compas_viewer.controller.Controller`
@@ -74,6 +79,11 @@ class Viewer:
         from compas_viewer import Viewer
         viewer = Viewer()
         viewer.show()
+
+    See Also
+    --------
+    :class:`compas.scene.Scene`
+
     """
 
     def __init__(
@@ -87,7 +97,7 @@ class Viewer:
         show_grid: Optional[bool] = None,
         configpath: Optional[str] = None,
     ):
-
+        super(Viewer, self).__init__()
         self.started = False
 
         # Custom or default config
@@ -122,13 +132,24 @@ class Viewer:
         self.app = QCoreApplication.instance() or QApplication(sys.argv)
         self.window = QMainWindow()
 
-        # Scene
-        self.scene = Scene(self, config=self.scene_config)
-
         # Controller
         self.controller = Controller(self, self.controller_config)
 
+        #  Selection
+        self.instance_colors: dict[tuple[int, int, int], ViewerSceneObject] = {}
+        self._instance_colors_generator = instance_colors_generator()
+
         # Render
+        self.grid = FrameObject(
+            Frame.worldXY(),
+            framesize=self.renderer_config.gridsize,
+            show_framez=self.renderer_config.show_gridz,
+            viewer=self,
+            is_selected=False,
+            is_locked=True,
+            is_visible=True,
+            config=self.scene_config,
+        )
         self.renderer = Renderer(self, self.renderer_config)
 
         # Layout
@@ -138,6 +159,14 @@ class Viewer:
         # `on` function
         self.timer: Timer
         self.frame_count: int = 0
+
+        #  Primitive
+        self.objects: list[ViewerSceneObject]
+
+    def __new__(cls, *args, **kwargs):
+        instance = super(Viewer, cls).__new__(cls)
+        Scene.viewerinstance = instance  # type: ignore
+        return instance
 
     # ==========================================================================
     # Scene
@@ -227,6 +256,117 @@ class Viewer:
         return outer
 
     # ==========================================================================
+    # Scene
+    # ==========================================================================
+
+    def add(
+        self,
+        item: Data,
+        parent: Optional[ViewerSceneObject] = None,
+        is_selected: bool = False,
+        is_locked: bool = False,
+        is_visible: bool = True,
+        show_points: Optional[bool] = None,
+        show_lines: Optional[bool] = None,
+        show_faces: Optional[bool] = None,
+        pointscolor: Optional[Union[Color, dict[Any, Color]]] = None,
+        linescolor: Optional[Union[Color, dict[Any, Color]]] = None,
+        facescolor: Optional[Union[Color, dict[Any, Color]]] = None,
+        lineswidth: Optional[float] = None,
+        pointssize: Optional[float] = None,
+        opacity: Optional[float] = None,
+        hide_coplanaredges: Optional[bool] = None,
+        use_vertexcolors: Optional[bool] = None,
+        **kwargs
+    ) -> ViewerSceneObject:
+        """
+        Add an item to the scene.
+        This function is inherent from :class:`compas.scene.Scene` with additional functionalities.
+
+        Parameters
+        ----------
+        item : Union[:class:`compas.geometry.Geometry`, :class:`compas.datastructures.Mesh`]
+            The geometry to add to the scene.
+        parent : :class:`compas_viewer.scene.ViewerSceneObject`, optional
+            The parent of the item.
+        is_selected : bool, optional
+            Whether the object is selected.
+            Default to False.
+        is_locked : bool, optional
+            Whether the object is locked (not selectable).
+            Default to False.
+        is_visible : bool, optional
+            Whether to show object.
+            Default to True.
+        show_points : bool, optional
+            Whether to show points/vertices of the object.
+            It will override the value in the scene config file.
+        show_lines : bool, optional
+            Whether to show lines/edges of the object.
+            It will override the value in the scene config file.
+        show_faces : bool, optional
+            Whether to show faces of the object.
+            It will override the value in the scene config file.
+        pointscolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
+            The color or the dict of colors of the points.
+            It will override the value in the scene config file.
+        linescolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
+            The color or the dict of colors of the lines.
+            It will override the value in the scene config file.
+        facescolor : Union[:class:`compas.colors.Color`, dict[Any, :class:`compas.colors.Color`], optional
+            The color or the dict of colors the faces.
+            It will override the value in the scene config file.
+        lineswidth : float, optional
+            The line width to be drawn on screen
+            It will override the value in the scene config file.
+        pointssize : float, optional
+            The point size to be drawn on screen
+            It will override the value in the scene config file.
+        opacity : float, optional
+            The opacity of the object.
+            It will override the value in the scene config file.
+        hide_coplanaredges : bool, optional
+            Whether to hide the coplanar edges of the mesh.
+            It will override the value in the scene config file.
+        use_vertexcolors : bool, optional
+            Whether to use vertex color.
+            It will override the value in the scene config file.
+        **kwargs : dict, optional
+            The other possible parameters to be passed to the object.
+
+        Returns
+        -------
+        :class:`compas.scene.SceneObject`
+            The scene object.
+        """
+
+        sceneobject: ViewerSceneObject = super(Viewer, self).add(  # type: ignore
+            item=item,
+            parent=parent,
+            viewer=self,
+            is_selected=is_selected,
+            is_visible=is_visible,
+            is_locked=is_locked,
+            show_points=show_points,
+            show_lines=show_lines,
+            show_faces=show_faces,
+            pointscolor=pointscolor,
+            linescolor=linescolor,
+            facescolor=facescolor,
+            lineswidth=lineswidth,
+            pointssize=pointssize,
+            opacity=opacity,
+            hide_coplanaredges=hide_coplanaredges,
+            use_vertexcolors=use_vertexcolors,
+            config=self.scene_config,
+            **kwargs
+        )
+
+        self.instance_colors[sceneobject.instance_color.rgb255] = sceneobject
+
+        return sceneobject
+
+    # ==========================================================================
     # Action
     # ==========================================================================
 
@@ -269,7 +409,7 @@ class Viewer:
             from compas.geometry import Transformation
             from compas_viewer import Viewer
             viewer = Viewer()
-            faces = viewer.scene.add(Mesh.from_obj(compas.get("faces.obj")))
+            faces = viewer.add(Mesh.from_obj(compas.get("faces.obj")))
             faces.transformation = Transformation()
             def pressed_action():
                 faces.transformation *= Scale.from_factors([1.1, 1.1, 1.1], Frame.worldXY())
@@ -299,3 +439,12 @@ class Viewer:
         self.controller.actions[name] = action
 
         return action
+
+    def clear(self, guids: Optional[Union[list[str], list[ViewerSceneObject]]] = None):
+        """Clear the scene."""
+        if guids is None:
+            guids = self.objects
+
+        for obj in guids:
+            self.remove(obj)
+            del obj
